rename games to gamesController
